[Refactor] 등록하기 모달창에 위치한 태그 입력 폼을 글 내용 작성 폼 밑으로 이동시켜라
- 등록하기 모달창에 위치한 태그 입력 폼을 글 내용 작성 폼 밑으로 이동
