+++ src/components/write/PublishModalForm.test.tsx
@@ -69,21 +69,4 @@ describe('PublishModalForm', () => {
       });
     });
   });
-
-  describe('태그를 입력한다', () => {
-    it('changeFields 이벤트가 발생해야만 한다', () => {
-      renderPublishModalForm(WRITE_FIELDS_FIXTURE);
-
-      const input = screen.getByPlaceholderText('태그를 입력하세요');
-
-      fireEvent.change(input, { target: { value: 'test' } });
-
-      fireEvent.keyPress(input, { key: 'Enter', code: 13, charCode: 13 });
-
-      expect(handleChangeFields).toBeCalledWith({
-        name: 'tags',
-        value: ['test'],
-      });
-    });
-  });
 });
